Replace the "pivotal" pieces with a properly maintained library
The "pivotal" time functions working around the y2038 problem had
several issues: they consist in unmaintained code distributed under
an ambigous and probably non-free license requiring to adversise
a dead website (see #13).

This commit replaces the "pivotal" code with a library that implements
the same functions and is properly maintained and licensed (MIT/Expat).
As the license is different from the one of libmseed (GPLv3), the new
files have been put in the separate time64/ directory, which includes
a LICENSE file and a README pointing to the upstream repository.

With this new library in place all the tests in tests/ pass.

Fixes #13.