Makefile mods to work with CMake as ExternalProject
When mseed is added as an ExternalProject to CMake, the installation of
libraries fails silently. The comparison ifneq (wildcard $(LIB_SO))
fails for some reason. This commit creates a common LIB_SO set of
variables for Darwin and Linux so that the ifneq comparison during
install time can be removed.