Simplify byte swapping routines
Firstly, compilers are smart enough that the naive copy-and-bit shift
method is converted to a register load/store + `bswap` intrinsic.
Specialized versions only serve to confuse the reader and end up with
the same or worse compiled code.

Secondly, the cast to `uint32_t*` on a `float` array (in
`msr_decode_float32` via `ms_gswap4a`) appears to violate strict
aliasing rules, such that an optimizing compiler _ignores the byte swap
entirely_.