Update numsamples and samplecnt consistently (#33)
* Check for invalid whence argument to mst_addspan and mst_addmsr

Check that whence argument to mst_addspan and mst_addmsr is
valid, which prevents modifications to MSTrace.numsmaples when
the number of samples has not actually changed.

* Update samplecnt and dataquality when there are datasamples

In mst_addmsr and msr_addspan, update the MSTrace.dataquality
and MSTrace.samplecnt only when the source data buffer, which is
either MSRecord.datasample or a pointer to a datasamples buffer,
is non-null and has data samples.

* Update samplecnt by number of samples actually added

In mst_addmsr, update MSTrace.samplecnt by the number of samples
actually copied into the MSTrace.datasamples buffer, which is
MSRecord.numsamples. This change makes the behavior of mst_addmsr
consistent with the behavior of mst_addspan.